v3.0.1 - Fix bug in Util.round, mod SendableVersion to log exception in init method.

+++ src/main/java/Team4450/Lib/Util.java
@@ -567,7 +567,7 @@ public class Util
 	{
 		BigDecimal bd = new BigDecimal(number);
 		bd = bd.setScale(decimalPlaces, rounding);
-		return bd.floatValue();
+		return bd.doubleValue();
 	}
 	
 	/**
